Fixed some bugs

+++ controllers/communities.js
@@ -4,6 +4,12 @@ const getAllCommunityPosts = async (req, res) => {
     const communityId = req.params.communityId
 
     try {
+        const community = await db.Community.findByPk(communityId)
+
+        if (!community) {
+            throw new Error('Community not found')
+        }
+
         const posts = await db.Post.findAll({
             where: {
                 communityId,
@@ -22,7 +28,6 @@ const getAllCommunityPosts = async (req, res) => {
 const getCommunityPost = async (req, res) => {
     const communityId = req.params.communityId
     const postId = req.params.postId
-    let error = 'Something went wrong'
 
     try {
         const post = await db.Post.findOne({
@@ -33,15 +38,14 @@ const getCommunityPost = async (req, res) => {
         })
 
         if (!post) {
-            error = 'Post not found'
-            throw new Error(error)
+            throw new Error('Post not found')
         }
 
         res.send(post)
     } catch (e) {
         console.error('Error:', e.message)
         res.send({
-            error,
+            error: 'Something went wrong',
         })
     }
 }
@@ -52,9 +56,12 @@ const getAllCommunityMembers = async (req, res) => {
     try {
         const community = await db.Community.findByPk(communityId)
 
+        if (!community) {
+            throw new Error('Community not found')
+        }
+
         const memberIds = await community.getUsers()
 
-        console.log('memberIds', memberIds)
         res.send(memberIds)
     } catch (e) {
         console.error('Error:', e.message)
@@ -70,30 +77,43 @@ const getCommunityAdminsOrModerators = async (option, req, res) => {
 
     try {
         if (option === 'admins') {
-            criteria = '$Communities.UserCommunity.isCreator$'
+            criteria = {
+                isAdmin: 1,
+            }
         } else if (option === 'moderators') {
-            criteria = '$Communities.UserCommunity.isModerator$'
+            criteria = {
+                isModerator: 1,
+            }
         } else {
             throw new Error('Invalid option')
         }
 
-        const moderators = await db.User.findAll({
-            include: [
-                {
-                    model: db.Community,
-                    as: 'Communities',
-                    where: {
-                        '$Communities.UserCommunity.communityId$': communityId,
-                        [criteria]: 1,
-                    },
-                },
-            ],
+        const community = await db.Community.findByPk(communityId)
+
+        if (!community) {
+            throw new Error('Community does not exist')
+        }
+
+        var users = await db.UserCommunity.findAll({
+            attributes: ['userId'],
+            where: {
+                ...criteria,
+                communityId,
+            },
         })
 
-        if (moderators.length == 0) {
-            res.send('Database error: No moderator found')
+        users = users.map((user) => user.toJSON().userId)
+
+        if (users.length == 0) {
+            res.send('Database error: No admin/moderator found')
         } else {
-            res.send(moderators)
+            users = await Promise.all(
+                users.map(async (id) => {
+                    const u = await db.User.findByPk(id)
+                    return u
+                })
+            )
+            res.send(users)
         }
     } catch (e) {
         console.error('Error:', e.message)
@@ -107,6 +127,9 @@ const createCommunity = async (req, res) => {
     const body = req.body
     const userId = req.params.userId
 
+    // add the user as the creator of the community
+    body.userId = userId
+
     try {
         const user = await db.User.findByPk(userId)
 
@@ -117,7 +140,7 @@ const createCommunity = async (req, res) => {
         const createdCommunity = await db.Community.create(body)
         await createdCommunity.addUser(user, {
             through: {
-                isCreator: 1,
+                isAdmin: 1,
                 isModerator: 1,
             },
         })
@@ -137,6 +160,11 @@ const updateCommunity = async (req, res) => {
         updatedAt: new Date(),
     }
     const communityId = req.params.id
+
+    // cannot update the creator of a community
+    if (body.userId) {
+        delete body.userId
+    }
 
     try {
         await db.Community.update(body, {
@@ -159,6 +187,12 @@ const deleteCommunity = async (req, res) => {
     const communityId = req.params.id
 
     try {
+        const community = await db.Community.findByPk(communityId)
+
+        if (!community) {
+            throw new Error('Community not found')
+        }
+
         await db.Community.destroy({
             where: {
                 id: communityId,
